+++ python/libcasm/mapping/methods/__init__.py
@@ -1,12 +1,11 @@
 """Easy-to-use mapping methods"""
-from ._methods import (
-    map_lattices_without_reorientation,
-)
-
 from ._mapping_methods import (
     make_mapped_lattice,
     make_mapped_structure,
     map_atoms,
     map_lattices,
     map_structures,
+)
+from ._methods import (
+    map_lattices_without_reorientation,
 )