+++ python/src/mapping_info.cpp
@@ -10,8 +10,8 @@
 #include "casm/mapping/AtomMapping.hh"
 #include "casm/mapping/LatticeMapping.hh"
 #include "casm/mapping/StructureMapping.hh"
-#include "casm/mapping/lattice_cost.hh"
 #include "casm/mapping/io/json_io.hh"
+#include "casm/mapping/lattice_cost.hh"
 #include "pybind11_json/pybind11_json.hpp"
 
 #define STRINGIFY(x) #x
@@ -771,9 +771,7 @@ PYBIND11_MODULE(_mapping_info, m) {
       },
       "Pretty-print JSON to string.", py::arg("data"));
 
-  m.def(
-      "lattice_isotropic_strain_cost", &isotropic_strain_cost,
-      R"pbdoc(
+  m.def("isotropic_strain_cost", &isotropic_strain_cost, R"pbdoc(
       Return the isotropic strain cost for a lattice deformation.
 
       Parameters
@@ -784,13 +782,14 @@ PYBIND11_MODULE(_mapping_info, m) {
 
       Returns
       -------
-      float : The isotropic strain cost for the lattice deformation given by :math:`F`.
+      cost: float
+          The isotropic strain cost for the lattice deformation given by :math:`F`. See
+          :func:`~libcasm.mapping.methods.map_atoms` for the definition.
       )pbdoc",
-      py::arg("deformation_gradient"));
+        py::arg("deformation_gradient"));
 
-    m.def(
-      "lattice_symmetry_breaking_strain_cost", &symmetry_breaking_strain_cost,
-      R"pbdoc(
+  m.def("symmetry_breaking_strain_cost", &symmetry_breaking_strain_cost,
+        R"pbdoc(
       Return the symmetry-breaking strain cost for a lattice deformation.
 
       Parameters
@@ -798,14 +797,16 @@ PYBIND11_MODULE(_mapping_info, m) {
       deformation_gradient : array_like, shape=(3,3)
           The parent-to-child deformation gradient tensor, :math:`F`, a shape=(3,3)
           matrix.
-      lattice1_point_group : list[:class:`xtal::SymOp`]
-          The point group of the parent crystal lattice.
+      lattice1_point_group : list[libcasm.xtal.SymOp]
+          The point group of the parent.
 
       Returns
       -------
-      float : The symmetry-breaking strain cost for the lattice deformation given by :math:`F`.
+      cost: float
+          The symmetry-breaking strain cost for the lattice deformation given by
+          :math:`F`. See :func:`~libcasm.mapping.methods.map_atoms` for the definition.
       )pbdoc",
-      py::arg("deformation_gradient"), py::arg("lattice1_point_group"));
+        py::arg("deformation_gradient"), py::arg("lattice1_point_group"));
 
 #ifdef VERSION_INFO
   m.attr("__version__") = MACRO_STRINGIFY(VERSION_INFO);